requirements.txt
pillow==11.3.0
pygame==2.6.1
